refactor: Marketplace.vue replaced to marketplace/
