+++ src/presentation/pages/marketplace/Marketplace.vue
@@ -17,8 +17,8 @@
 
 <script setup lang="ts">
 import useMarketplace from '@/application/services/useMarketplace';
-import EditorToolElement from '../components/marketplace/EditorToolElement.vue';
-import Button from '../components/button/Button.vue';
+import EditorToolElement from '@/presentation/components/marketplace/EditorToolElement.vue';
+import Button from '@/presentation/components/button/Button.vue';
 
 const { tools } = useMarketplace();
 </script>
